// src/shared/domain/database-types.ts
export interface Statistics {
  totalPrompts: number;
  deletedPrompts: number;
  favoritePrompts: number;
  promptsWithImages: number;
  totalImages: number;
  deletedImages: number;
  favoriteImages: number;
  referencedImages: number;
}